Modules/Discord/Helix/requirements.txt
absl-py==1.0.0
aenum==3.1.5
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiomysql==0.0.22
aiosignal==1.2.0
aiosqlite==0.17.0
APScheduler==3.8.1
argon2-cffi==21.1.0
asgiref==3.4.1
astunparse==1.6.3
async-cse==0.3.0
async-generator==1.10
Async-OAuthlib==0.0.9
async-timeout==3.0.1
asyncio==3.4.3
asyncio-extras==1.3.2
asyncpraw==7.5.0
asyncprawcore==2.3.0
attrs==21.2.0
audioread==2.1.9
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.8
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
cycler==0.11.0
Cython>=0.28.5
dataclasses==0.6
DateTime==4.3
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.3
discord-ext-ipc==2.1.1
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@fbb8803779373357e274e1540b368365fd9d8074
discord.py==1.7.3
Django==3.2.9
django-classy-tags==2.0.0
django-cms==3.9.0
django-formtools==2.3
django-sekizai==2.0.0
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
flatbuffers==2.0
frozenlist==1.2.0
fsspec==0.8.5
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.3.3
geographiclib==1.52
geojson==2.3.0
geopy==2.2.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h11==0.12.0
h2==4.1.0
h5py==3.6.0
hpack==4.0.0
httptools==0.3.0
hypercorn==0.12.0
hyperframe==6.0.1
idna==2.10
importlib-metadata==4.8.2
ipykernel==6.5.1
ipython==7.30.0

ipython-genutils==0.2.0
ipywidgets==7.6.5
iso8601==0.1.16
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.11.2
joblib==1.0.0
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
librosa==0.7.2
llvmlite==0.37.0
Markdown==3.3.6
MarkupSafe==1.1.1
matplotlib==3.3.3
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.2.1
multidict==5.2.0
mysql-connector==2.2.9
mysql-connector-python==8.0.27
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.5
nose==1.3.7
notebook==6.4.6
numba==0.54.1
numpy==1.20.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==21.3
pandarallel==1.4.8
pandas==1.1.5
pandocfilters==1.5.0
parsedatetime==2.6
parso==0.8.2
pickleshare==0.7.5
Pillow==8.0.1
priority==2.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.23
protobuf==3.19.1
psutil==5.8.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycparser==2.21
pyDeprecate==0.3.1
pydub==0.23.1
Pygments==2.10.0
PyJWT==2.3.0
PyMySQL==0.9.3
PyNaCl==1.4.0
pyowm==2.9.0
pyparsing==3.0.6
pypika-tortoise==0.1.1
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.15.0
pytorch-lightning==1.1.1
pyttsx3==2.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywinpty
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.2
quart==0.16.1
Quart-Discord==2.1.4
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
rich==10.14.0
rsa==4.8
sacremoses==0.0.43
sanic==21.9.3
sanic-routing==0.7.2
scipy==1.7.3
seaborn==0.11.1
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.5.0
simplejson==3.17.6
six==1.16.0
sonopy==0.1.2
SoundFile==0.10.3.post1
SpeechRecognition==3.8.1
sqlparse==0.4.2
sympy==1.9
tabulate==0.8.7
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tokenizers==0.9.4
toml==0.10.2
torch==1.10.0
torchaudio==0.10.0
torchmetrics==0.6.0
torchtext==0.6.0
torchvision==0.11.1
tornado==6.1
tortoise-orm==0.17.8
tqdm==4.54.1
traitlets==5.1.1
transformers==4.1.1
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
update-checker==0.18.0
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.1
Werkzeug==2.0.0
widgetsnbextension==3.5.2
wrapt==1.13.3
wsproto==1.0.0
yarl==1.7.2
youtube-dl==2021.6.6
zipp==3.6.0
zope.interface==5.4.0
